+++ app/src/main/java/ru/practicum/android/diploma/core/data/db/AppDatabase.kt
@@ -10,7 +10,6 @@ import androidx.room.RoomDatabase
     version = 1,
     exportSchema = false
 )
-
 abstract class AppDatabase : RoomDatabase() {
 
     abstract fun favoriteVacancyDao(): FavoriteVacancyDao
