+++ app/src/main/java/ru/practicum/android/diploma/features/search/VacancySearchFragment.kt
@@ -1,11 +1,14 @@
-package ru.practicum.android.diploma.features.vacanciesFavorite
+package ru.practicum.android.diploma.features.search
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.Scaffold
@@ -13,12 +16,13 @@ import androidx.compose.material3.Text
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
+import androidx.compose.ui.unit.dp
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
 import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.core.theme.AppTheme
 
-class VacancyFavoriteFragment : Fragment() {
+class VacancySearchFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -40,11 +44,20 @@ class VacancyFavoriteFragment : Fragment() {
                             Box(
                                 modifier = Modifier.fillMaxSize()
                             ) {
-                                Text("VacancyFavoriteFragment")
-                                Button(onClick = {
-                                    findNavController().navigate(R.id.action_to_detail)
-                                }) {
-                                    Text("Избранное детальное")
+                                Column() {
+                                    Text("SEARCH")
+                                    Button(onClick = {
+                                        findNavController().navigate(R.id.action_to_detail)
+                                    }) {
+                                        Text("Поисковой элемент - детальное")
+                                    }
+                                    Spacer(modifier = Modifier.height(12.dp))
+
+                                    Button(onClick = {
+                                        findNavController().navigate(R.id.action_to_filter)
+                                    }) {
+                                        Text("Фильтр")
+                                    }
                                 }
                             }
                         }
