+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,12 +6,12 @@
 
     <fragment
         android:id="@+id/searchFragment"
-        android:name="ru.practicum.android.diploma.features.vacanciesSearch.VacancySearchFragment"
+        android:name="ru.practicum.android.diploma.features.search.VacancySearchFragment"
         android:label="VacancySearchFragment" />
 
     <fragment
         android:id="@+id/favoriteFragment"
-        android:name="ru.practicum.android.diploma.features.vacanciesFavorite.VacancyFavoriteFragment"
+        android:name="ru.practicum.android.diploma.features.favorite.VacancyFavoriteFragment"
         android:label="VacancyFavoriteFragment" />
 
     <fragment
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/vacancyDetailFragment"
-        android:name="ru.practicum.android.diploma.features.vacancyDetail.VacancyDetailFragment" />
+        android:name="ru.practicum.android.diploma.features.detail.VacancyDetailFragment" />
 
     <fragment
         android:id="@+id/settingFilterFragment"
